+++ changelog.md
@@ -12,6 +12,7 @@
 - DriverInterface::handleImageInput()
 - DriverInterface::handleColorInput()
 - DataUri::class
+- Default resolution is now 72 ppi for new images
 
 ## API Changes